requirements.txt
certifi==2020.11.8
chardet==3.0.4
et-xmlfile==1.0.1
idna==2.10
jdcal==1.4.1
openpyxl==3.0.5
python-dateutil==2.8.1
python-gitlab==2.5.0
requests==2.25.0
six==1.15.0
urllib3==1.26.2
